refactor speech synthesis model
